pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scripts*"]

[project]
name = "stables-analytics"
version = "0.1.0"
description = "Stablecoins Analytics Pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.6.1",
    "dbt-core>=1.10.13",
    "dbt-postgres>=1.9.1",
    "dbt-snowflake>=1.10.2",
    "dlt[postgres]>=1.17.0",
    "hypersync>=0.8.5",
    "jupyter>=1.1.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.22",
    "pandas>=2.3.3",
    "polars>=1.33.1",
    # psycopg is provided by dlt[postgres] and dbt-postgres (psycopg2-binary)
    # Explicitly installing psycopg v3 causes conflicts with dbt-postgres
    "pyarrow>=21.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "rich>=13.6.0",
    "snowflake-connector-python>=3.17.4",
    "sqlalchemy>=2.0.43",
    "langfuse>=2.59.5",
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "langchain-core>=1.2.7",
    "langgraph>=1.0.7",
    "langchain-tavily>=0.2.17",
    "langchain-experimental>=0.4.1",
    "qdrant-client>=1.16.2",
    "psycopg>=3.3.2",
    "matplotlib>=3.10.8",
    "snowflake-core>=1.11.0",
    "snowflake-snowpark-python>=1.44.0",
]

[project.scripts]
kafka-init-schema = "scripts.kafka.init_kafka_schema:init_schema"
kafka-producer = "scripts.kafka.producer:main"
kafka-consumer = "scripts.kafka.consumer:main"

[dependency-groups]
dev = ["sqlfluff>=3.5.0", "sqlfluff-templater-dbt>=3.5.0"]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
